requirements.txt
asyncio~=3.4.3
loguru~=0.7.0
python-dotenv~=1.0.0
aiohttp~=3.8.4
aptos-sdk~=0.7.1
